Vision/requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
httpx==0.27.0

# ── Reconocimiento facial ─────────────────────────────────────────────────────
dlib==20.0.0
face_recognition==1.3.0

# ── Procesamiento de imágenes ─────────────────────────────────────────────────
numpy==1.26.4
Pillow==11.1.0

# ── Configuración ─────────────────────────────────────────────────────────────
python-dotenv==1.0.1